Доблавил тесты к компоненту Host и к reducers
