+++ src/reducer/data/data.test.js
@@ -1,0 +1,96 @@
+import {reducer, ActionType} from "./data.js";
+
+
+const offerWithPremium = {
+  bedroom: 2,
+  coords: [48.865610000000004, 2.350499],
+  description: `Discover daily local life in city center.`,
+  facilities: [`Air conditioning`, `Breakfast`],
+  grade: 3.6,
+  host: {
+    avatarUrl: `img/avatar-angelina.jpg`,
+    id: 25,
+    isPro: true,
+    name: `Angelina`,
+    // id: 0,
+  },
+  isFavourite: false,
+  isPremium: true,
+  locationZoom: 16,
+  maxAdults: 8,
+  pictures: [`https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/3.jpg`, `https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/12.jpg`],
+  previewImage: `https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/5.jpg`,
+  price: 397,
+  reviews: {
+    image: `img/avatar-max.jpg`,
+    text: `A quiet cozy and picturesque.`,
+    name: `Max`,
+    time: `April 2019`
+  },
+  title: `Penthouse, 4-5 rooms + 5 balconies`,
+  type: `hotel`,
+};
+
+const offerWithFavourite = {
+  bedroom: 2,
+  coords: [48.865610000000004, 2.350499],
+  description: `Discover daily local life in city center.`,
+  facilities: [`Air conditioning`, `Breakfast`],
+  grade: 3.6,
+  host: {
+    avatarUrl: `img/avatarangelina.jpg`,
+    id: 25,
+    isPro: false,
+    name: `Angela`,
+    // id: 2,
+  },
+  isFavourite: true,
+  isPremium: true,
+  locationZoom: 16,
+  maxAdults: 8,
+  pictures: [`https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/3.jpg`, `https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/12.jpg`],
+  previewImage: `https://htmlacademy-react-3.appspot.com/six-cities/static/hotel/5.jpg`,
+  price: 397,
+  reviews: {
+    image: `img/avatar-max.jpg`,
+    text: `A quiet cozy and picturesque.`,
+    name: `Max`,
+    time: `April 2019`
+  },
+  title: `Penthouse, 4-5 rooms + 5 balconies`,
+  type: `hotel`,
+};
+
+const offers = [
+  {
+    city: `Paris`,
+    cityCoords: [48.85661, 2.351499],
+    cityZoom: 13,
+    localOffers: [offerWithFavourite],
+  },
+  {
+    city: `Amsterdam`,
+    cityCoords: [52.38333, 4.9],
+    localOffers: [
+      offerWithPremium,
+    ]
+  }];
+
+describe(`Test of reduser data.js`, () => {
+  it(`Reducer without additional parameters should return initial state`, () => {
+    expect(reducer(void 0, {})).toEqual({
+      offers: [],
+    });
+  });
+
+  it(`Reducer should update offers by load offers`, () => {
+    expect(reducer({
+      offers: [],
+    }, {
+      type: ActionType.LOAD_OFFERS,
+      payload: offers,
+    })).toEqual({
+      offers,
+    });
+  });
+});
